services/core/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
psycopg2-binary==2.9.9
redis==5.2.0
clickhouse-connect==0.8.2
pika==1.3.2
openai==1.54.3
requests==2.32.3
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==9.0.0
httpx==0.27.2
